feat(singup): handle signature and view
Update the user interface for handle selection and account choosing to align with the provided design mockups.

Update the radio buttons for selections with a dropdown menu.

Disable the "Next Page" button and only enable it when the account selection and handle claim have been validated as correct.

Navigate the user to the handle signing review page for successful validation of the handle claim.

issue-95