fix: simplify async stores (#118)
# Description
This PR simplifies derived stores that need to be populated
asynchronously. Previously the stores themselves returned a Promise,
based on [this Stackoverflow
response](https://stackoverflow.com/a/73873614/6188946). This required
consumers of the store to be async as well. However, there is a simpler
way, using the alternate Svelte `derived` method that includes a `set`
method parameter (note, mention of this method is conspicuously missing
from the Svelte tutorial, though it is in the API docs).

This small refactor has a large benefit in that page and component code
that consumes these stores no longer needs to be async.

Closes #117