fix: do not connect to chain until signin request received from client (#145)
# Description
This PR fixes the issue where the MSAs in a wallet are not detected
correctly at first

Closes #144 

## Details
The issue was due to the fact that at app initialization, we were at
first connecting to the "default" chain (localhost), and so initially
there were MSAs detected on the local chain. Then the chain connection
was switched to the node/chain desired by the client (ie, rococo or
mainnet), which contained no MSAs. But the check for MSAs present had
already happened, and thus the user was directed incorrectly to the
sign-in flow, which dynamically update to (correctly) show no MSAs
available to sign in.

# Solution
Eliminated the concept of a "default" chain node; instead, obtaining a
connection to the chain is a Promise that does not resolve until a node
URL is set after receiving a signin request from the client.

Added benefit: we now no longer need to specify the chain URL as a route
query parameter; it can be part of the signin request payload as
originally intended.