Fixing V1 (#172)
There is a v2 coming of SIWF. This prepares to leave the SIWF v1 working
and able to be patched, but moves the deployment such that we can deploy
multiple versions. All without breaking the existing deployment.

## Testing

Test using the [SAT](https://github.com/AmplicaLabs/social-app-template)
with both the versionless and versioned urls.